Swapping (call-by-reference)
Swapping call-by-value: When a function is called by value, the arguments take the copy of the value into the function. 
Call-by-reference: when a function is called by reference, the pointer to that variable is passed and any changes in the value gets affected as it is the same memory location.